Time: 5 ms (34.50%) | Memory: 7 MB (61.35%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool judgeSquareSum(int c) {
-        for(long a = 0; a * a <=c; a++){
+        for(double a=0; a*a <=c; a++){
             double b = sqrt(c - a * a);
             if(b == (int)b){
                 return true;
